requirements.txt
attrs==19.3.0
black==19.10b0
demoji==0.2.1
emoji==0.5.4
fasttext==0.9.2
flair==0.4.5
gensim==3.8.3
google-api-python-client==1.8.3
matplotlib==3.2.1
nltk==3.5
numpy==1.18.4
pandas==1.0.3
scikit-learn==0.23.0
torch==1.5.0
transformers==2.9.1
tensorflow==2.2.0
lime==0.2.0.0
fast-bert==1.7.1
eli5==0.10.1
seaborn==0.10.1